requirements.txt
fastmcp>=0.4.0
openai-agents>=0.0.12
python-dotenv>=1.0.1
httpx>=0.27.0
psycopg2-binary>=2.9.0
fastapi>=0.115.0
uvicorn>=0.32.0

